First version of analysis
Eye data files are read out and processed with Matlab; resulting data
frame is analyzed with R.

Both are subject to change: in matlab, the parameters to determine
which trials to include might not be optimal (and for flexibility
trials should only be averaged later); the R-file is just a quick and
dirty stab at a simple analysis. But it's enough to look at the general
effects as published in Kanai et al. (2012).