Update task with slight changes made in lab
The final version of the task was debugged with the eye tracker in
the lab before starting the experiment. Some slight changes were made
then that were not updated on the server. These mostly involve comments,
line breaks and such. The one exception is changing the name of a
variable (incorrect in the old version: tFixOnset) in the Main_ET.m file
(correct in the new version: tISIOnset(1)).