Fix starting task at different leg/block
Fixes a bug so that when starting the task at a different point, it
will complete the whole task from that point onwards.

It was already possible to start the task at a point different from the
beginning, in case it ever needed to be restarted (e.g. a crash).
However, this did not work properly: if you wanted to start at leg
2 block 3, it would thereafter move to leg 3 block 3, instead of
starting at block 1 again for leg 3.