Print stage of task to command window
The screen in the lab was taken away for awhile; when it returned, the
subject screen and the experimenter screen were no longer mirrored. Thus
the experimenter could not see what the subject was doing at a given
moment (in a break, doing the task, etc.). Now short messages are
printed to the command window with this information.

N.B. the change in sacctDCS_setup is non-substantial (default screen
distance was not updated on the server) and has nothing to do with this.