Add tDCS code mapping for subject 16 onwards
After subject 15, the tDCS device was reprogrammed again, such that
anodal and cathodal stimulation were assigned different letter codes.
Here the script that parses the eye data txt files is adapted, such that
the data from each sessions is correctly assigned to "anodal" or
"cathodal".

N.B.: This commit was made much later than the changes to the file,
which predate presentation of this data at VSS in May 2017.

This change was made before end of data collection, but the
experimenter that ran the subjects was not the same as the one that
adapted the script, so the study was still double-blind. At a much
later time point, the experimenter that changed the script also ran
some subjects, but by then he had forgotten the mapping again (and did
not look at the script until after data collection).