Add x- and y-drift data columns
For each trial also writes out the difference between measured fixation
coordinates and actual fixation location at the end of each break. This
can be used to do a drift correction: if we trust that people are
actually fixating at this point, this offset can subtracted from the
saccade endpoints to correct for any drifts that may have occured since
calibration.