MRI coordinates in standard space
First commit of a script ("mni_coords.sh") that was used previously to
compute the coordinates of the FEF in MNI space ("FEF_MNIcoords.txt").

The average coordinate was used to make a figure for poster
presentations (e.g. for VSS 2017), but the script and file were not
committed at the time.

The present version incorporated subjects 1-20; they will later be
updated for the full dataset.