Fix indexing bug for sorted Bayes Factors
Function would look at the last row of data frame with Bayes Factors to
extract all effect names). For unsorted Bayes Factor tables, this last
row has the fullest model, but for sorted tables, this is not
necessarily true. Fixed so it now extracts effects from the row with
the longest name (i.e. the row with the model that includes all
effects).