Modify script to read coords from and write to csv
Previously, to get the MNI coordinates of the FEF for each subject: (1)
a text file with native coordinates for each subject was read in, and
(2) a text file with all the MNI coordinates was written out.

Now both the in- and outputs are csv files with the same column
structure.