Force scan orientation to match MNI template
Sometimes the images are still oriented as acquired by the scanner, not
in the default orientation matching the template (e.g. the saggital
slice with anterior on the right, posterior on the left). In this case
the coordinate translation will be off, just due to the mismatch in
orientations. In MRIcron and on the neuronav, scans are also always
displayed in the standard orientation.